+++ py4DSTEM/classes/__init__.py
@@ -1,6 +1,5 @@
 _emd_hook = True
 
 from py4DSTEM.classes.virtualdiffraction import VirtualDiffraction
-from py4DSTEM.classes.virtualimage import VirtualImage
