+++ py4DSTEM/data/__init__.py
@@ -2,7 +2,6 @@ _emd_hook = True
 
 from py4DSTEM.data.calibration import Calibration
 from py4DSTEM.data.data import Data
-from py4DSTEM.data.datacube import DataCube
 from py4DSTEM.data.diffractionslice import DiffractionSlice
 from py4DSTEM.data.realslice import RealSlice
 from py4DSTEM.data.qpoints import QPoints
