+++ py4DSTEM/datacube/__init__.py
@@ -1,0 +1,6 @@
+_emd_hook = True
+
+from py4DSTEM.datacube.datacube import DataCube
+from py4DSTEM.datacube.virtualimage import VirtualImage
+
+
