+++ py4DSTEM/process/__init__.py
@@ -2,7 +2,6 @@ from py4DSTEM.process.polar import PolarDatacube
 
 from py4DSTEM.process import diskdetection
 from py4DSTEM.process import virtualdiffraction
-from py4DSTEM.process import virtualimage
 from py4DSTEM.process import latticevectors
 from py4DSTEM.process import phase
 from py4DSTEM.process import calibration
