+++ py4DSTEM/process/diskdetection/probe_gen_methods.py
@@ -5,7 +5,7 @@ from scipy.ndimage import (
     binary_opening, binary_dilation, distance_transform_edt)
 
 from emdfile import tqdmnd
-from py4DSTEM.data import DataCube
+from py4DSTEM.datacube import DataCube
 from py4DSTEM.process.diskdetection import Probe
 from py4DSTEM.process.utils import get_shifted_ar, get_shift
