+++ py4DSTEM/process/phase/iterative_dpc.py
@@ -17,7 +17,8 @@ except ImportError:
     cp = None
 
 from emdfile import Array, Custom, Metadata, _read_metadata, tqdmnd
-from py4DSTEM.data import Calibration, DataCube
+from py4DSTEM.data import Calibration
+from py4DSTEM.datacube import DataCube
 from py4DSTEM.process.phase.iterative_base_class import PhaseReconstruction
 
 warnings.simplefilter(action="always", category=UserWarning)
