+++ py4DSTEM/visualize/__init__.py
@@ -3,5 +3,4 @@ from py4DSTEM.visualize.show import *
 from py4DSTEM.visualize.vis_RQ import *
 from py4DSTEM.visualize.vis_grid import *
 from py4DSTEM.visualize.vis_special import *
-from py4DSTEM.visualize.virtualimage import *
